Move alarm info setting out to seperate file
